+++ R/ggraphgrid.R
@@ -18,11 +18,6 @@
 #' Application of Approximate Pattern Matching in Two Dimensional
 #' Spaces to Grid Layout for Biochemical Network Maps.
 #' PLoS One 7 (6): e37739. doi: https://doi.org/10.1371/journal.pone.0037739.
-#' @examples
-#' ggraphgrid(mpn_elite_mex)
-#' ggraphgrid(mpn_elite_usa_advice)
-#' ggraphgrid(mpn_elite_usa_money)
-#' ggraphgrid(mpn_ryanair)
 #' @export
 ggraphgrid <- function(x, algorithm = c("kk", "fr")) {
   .Deprecated("autographr(x, 'frgrid'")